requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.45.1
transformers==4.52.3
torch==2.7.0
python-dotenv==1.0.1
pydantic==2.6.3
pytest==8.0.2
black==24.2.0
isort==5.13.2
flake8==7.0.0
python-multipart==0.0.9
sqlalchemy==2.0.0
psycopg2-binary==2.9.9
alembic==1.13.1 